Add full Y/A/D/N flags to existing functions
Also includes testing for a new build_segment_dictionaries function for users to allow for segments to be specified. This skirts an issue where opdata exists in CICS and NETVIEW segments. While this is not a problem elsewhere, it could be good to add the option to specify segments for ease of processing on some internal calls.